requirements.txt
numpy
opencv-python
ultralytics
python-dotenv